customer-service/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
pymongo==4.16.0
pydantic==2.12.5
pydantic[email]==2.12.5
python-dotenv==1.2.2
passlib[bcrypt]==1.7.4
httpx==0.28.1
